templates (el8): clean up %build and %install sections
We don't need complicated overrides in kmod-%{kmod_name}.conf, so
we can strip this back to the same as we had in el7 as we simply
want to override all kernels (Red Hat only want override the current
kernel series in their kmods as they are intended to be driver updates
that will be incorporated into the next kernel point release update).

Clean up make statement. Remove the EXTRA_CFLAGS as these should be
picked up as default from the kernel we are building against.
Also add %{?_smp_mflags} so we can use all available CPUs.

Clean up %install. Lets just install everything manually as we did for el7.

Strip the modules - we manually strip the modules at the end of install,
rather than allowing strip-to-file to strip them as this also strips
the SB signature.

Signed-off-by: Philip J Perry <phil@elrepo.org>